\name{conservation}

\docType{data}

\alias{conservation}

\title{Conservation of liquid}

\description{
	
	This data set is from a conservation of liquid task adminstered to 101
	children aged 6-10, measured at 5 occasions.  The data provided here
	are those analyzed in \cite{Schmittmann et al.  2005} in Illustration
	1, which is part of a larger data set described \cite{Van der Maas,
	1993}.
	
}

\usage{
data(discrimination)
}

\format{
A data frame with 101 rows with 6 variables:
  \describe{
	\item{\code{sex}}{a factor giving the sex of the participants.}
	\item{\code{r1}}{participant responses at measurement occascion 1.}
	\item{\code{r2}}{participant responses at measurement occascion 2.}
	\item{\code{r3}}{participant responses at measurement occascion 3.}
	\item{\code{r4}}{participant responses at measurement occascion 4.}
	\item{\code{r5}}{participant responses at measurement occascion 5.}
  }
}

\source{

	van der Maas, H. L. J..  (1993).  \emph{Catastrophe analysis of
	stagewise cognitive development: Model, method and applications.}
	Doctoral dissertation (Dissertatie reeks 1993-2), University of
	Amsterdam.
	
	Schmittmann, V. D., Dolan, C. V., van der Maas, H. L., & Neale, M. C.
	(2005).  Discrete latent Markov models for normally distributed response
	data.  \emph{Multivariate Behavioral Research}, 40(4), 461-488.
	
}

\keyword{datasets}
