\name{disc42}

\docType{data}

\alias{disc42}

\title{Discrimination Learning Data}

\description{
	This data set is a selection of six participants' responses from the 
	\code{discrimination} data set. 
}

\usage{
data(disc42)
}

\format{
A data frame consisting of the following variable:
  \describe{
	\item{\code{acc}}{a factor of accuracy scores (incorrect/correct)}
  }
}

\source{
	Maartje E. J. Raijmakers, Conor V. Dolan and Peter C. M. Molenaar 
	(2001).  Finite mixture distribution models of simple discrimination
	learning. \emph{Memory & Cognition}, vol 29(5).

	Ingmar Visser, Verena D. Schmittmann, and Maartje E. J. Raijmakers
	(2007).  Markov process models for discrimination learning.  In: Kees
	van Montfort, Han Oud, and Albert Satorra (Eds.), \emph{Longitudinal
	models in the behavioral and related sciences}, Mahwah (NJ): Lawrence
	Erlbaum Associates.

	Verena D. Schmittmann, Ingmar Visser and Maartje E. J. Raijmakers
	(2006).  Multiple learning modes in the development of rule-based
	category-learning task performance.  \emph{Neuropsychologia, vol
	44(11)}, p.  2079-2091.
}

\keyword{datasets}
