\name{balance8}

\docType{data}

\alias{balance8}

\title{Balance scale data}

\description{

	Data are from 8 repeated measurements of 25 balance scale items with 1004 
	participants ranging from 6-17 years of age. The data are aggregated over
	5 item types: weight (w), distance (d), conflict weight (cw), conflict 
	distance (cd), and conflict balance (cb). 
	
}

\usage{data(balance8)}
\format{
  A data frame with 8032 observations on the following 23 variables.
  \describe{
    \item{\code{id}}{a numeric vector}
    \item{\code{age}}{a numeric vector}
    \item{\code{sex}}{a factor with levels \code{m} \code{v}}
    \item{\code{class}}{a numeric vector}
    \item{\code{school}}{a factor with levels \code{primary} \code{secondary}}
    \item{\code{time}}{a numeric vector}
    \item{\code{missing}}{a factor with levels \code{Not Selected} \code{Selected}}
    \item{\code{wc}}{a numeric vector}
    \item{\code{wi}}{a numeric vector}
    \item{\code{ws}}{a numeric vector}
    \item{\code{dc}}{a numeric vector}
    \item{\code{di}}{a numeric vector}
    \item{\code{ds}}{a numeric vector}
    \item{\code{cwc}}{a numeric vector}
    \item{\code{cwi}}{a numeric vector}
    \item{\code{cws}}{a numeric vector}
    \item{\code{cdc}}{a numeric vector}
    \item{\code{cdi}}{a numeric vector}
    \item{\code{cds}}{a numeric vector}
    \item{\code{cbc}}{a numeric vector}
    \item{\code{cbi}}{a numeric vector}
    \item{\code{cbs}}{a numeric vector}
    \item{\code{trials}}{a numeric vector}
  }
}

\source{
	
	Brenda Jansen & Ingmar Visser. Manuscript in preparation.
	Corresponding author: i.visser@uva.nl
	
}

\examples{
data(balance8)
}

\keyword{datasets}
