\name{dccs}

\docType{data}

\alias{dccs}

\title{Dimensional Change Card Sort Task Data}

\description{
	This data set contains . 
}

\usage{
data(dccs)
}

\format{
A data frame consisting of the following variable:
  \describe{
	\item{\code{pp}}{}
  \item{\code{ageM}}{(numeric) age in months}
  \item{\code{ageY}}{(numeric) age in years}
  \item{\code{ageGr}}{(numeric) age group}
  \item{\code{sex}}{(numeric) age group}
  \item{\code{nTrPre}}{(numeric) number of trials completed pre-switch}
  \item{\code{nTrPre}}{(numeric) number of trials correct in pre-switch trials}
  \item{\code{t1-post} - \code{t6post}}{(numeric) indicator variable for each post switch trial indicating correct (1) or incorrect (0)}
  \item{\code{nCorPost}}{(numeric) number of trials correct post-switch}
  \item{\code{passPost}}{(numeric) indicator variable whether post-switch test was passed}
  }
}

\source{
	Bianca M.C.W. van Bers, Ingmar Visser, Tessa J.P. van Schijndel, Dorothy J. Mandell and Maartje E.J. Raijmakers (2011). The dynamics of development on the Dimensional Change Card Sorting task. \emph{Developmental Science, vol 14(5)}, 960-971.
}

\keyword{datasets}
